feat : 인수테스트 템플릿 추가
