+++ src/test/java/likelion/festival/notification/ui/NotificationControllerTest.java
@@ -9,15 +9,12 @@ import likelion.festival.notification.application.dto.NotificationUpdateDto;
 import likelion.festival.notification.domain.Notification;
 import likelion.festival.notification.domain.NotificationType;
 import likelion.festival.notification.domain.repository.NotificationRepository;
+import likelion.festival.support.AcceptanceTest;
 import likelion.festival.support.fixture.NotificationFixtureGenerator;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 
 import java.util.List;
@@ -26,27 +23,13 @@ import java.util.stream.Collectors;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class NotificationControllerTest {
+class NotificationControllerTest extends AcceptanceTest {
 
     @Autowired
     private NotificationFixtureGenerator notificationFixtureGenerator;
 
     @Autowired
     private NotificationRepository notificationRepository;
-
-    @LocalServerPort
-    private int port;
-
-    @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-    }
-
-    @AfterEach
-    void cleanUp() {
-        notificationRepository.deleteAll();
-    }
 
     @DisplayName("공지사항 생성")
     @Nested
